Float

+++ src/engine/geom/shape/Rectangle.java
@@ -23,8 +23,8 @@ public class Rectangle extends Polygon {
      * @param width the width of the rectangle
      * @param height the height of the rectangle
      */
-    public Rectangle(double x, double y, double width, double height) {
-        super(new double[] {x, x + width, x + width, x}, new double[] {y, y, y + height, y + height});
+    public Rectangle(float x, float y, float width, float height) {
+        super(new float[] {x, x + width, x + width, x}, new float[] {y, y, y + height, y + height});
         size = new Dimension(width, height);
     }
     
@@ -34,7 +34,7 @@ public class Rectangle extends Polygon {
      * @return the x coordinate
      * @see setX
      */
-    public double getX() {
+    public float getX() {
         return points[0].x;
     }
 
@@ -44,7 +44,7 @@ public class Rectangle extends Polygon {
      * @return the y coordinate
      * @see setY
      */
-    public double getY() {
+    public float getY() {
         return points[0].y;
     }
 
@@ -54,7 +54,7 @@ public class Rectangle extends Polygon {
      * @return the width of the dimension
      * @see setWidth
      */
-    public double getWidth() {
+    public float getWidth() {
         return size.getWidth();
     }
 
@@ -64,7 +64,7 @@ public class Rectangle extends Polygon {
      * @return the height of the dimension
      * @see setHeight
      */
-    public double getHeight() {
+    public float getHeight() {
         return size.getHeight();
     }
 
@@ -83,8 +83,8 @@ public class Rectangle extends Polygon {
      * @return the coordinates
      * @see setLocation
      */
-    public Point2D.Double getLocation() {
-        return new Point2D.Double(getX(), getY());
+    public Point2D.Float getLocation() {
+        return new Point2D.Float(getX(), getY());
     }
 
     /**
@@ -102,8 +102,8 @@ public class Rectangle extends Polygon {
      * 
      * @return the rotation theta
      */
-    public static double getRotation(double x, double y, double oppositeX, double oppositeY) {
-        return Math.atan((x - oppositeX) / (y - oppositeY));
+    public static float getRotation(float x, float y, float oppositeX, float oppositeY) {
+        return (float) Math.atan((x - oppositeX) / (y - oppositeY));
     }
 
     /**
@@ -112,9 +112,9 @@ public class Rectangle extends Polygon {
      * @param x the new x coordinate
      * @see getX
      */
-    public void setX(double x) {
+    public void setX(float x) {
         validate(x, "x");
-        double offsetX = Math.abs(x - points[0].x);
+        float offsetX = Math.abs(x - points[0].x);
         translateX(offsetX);
     }
 
@@ -124,9 +124,9 @@ public class Rectangle extends Polygon {
      * @param y the new x coordinate
      * @see getY
      */
-    public void setY(double y) {
+    public void setY(float y) {
         validate(y, "y");
-        double offsetY = Math.abs(y - points[0].y);
+        float offsetY = Math.abs(y - points[0].y);
         translateY(offsetY);
     }
 
@@ -138,8 +138,8 @@ public class Rectangle extends Polygon {
      * @throws IllegalArgumentException if the width is negative
      * @see getWidth
      */
-    public void setWidth(double width) {
-        double offset = Math.abs(width - getWidth());
+    public void setWidth(float width) {
+        float offset = Math.abs(width - getWidth());
         points[1].x += offset;
         points[2].x += offset;
         size.setWidth(width);
@@ -153,8 +153,8 @@ public class Rectangle extends Polygon {
      * @throws IllegalArgumentException if the height is negative
      * @see getHeight
      */
-    public void setHeight(double height) {
-        double offset = Math.abs(height - getHeight());
+    public void setHeight(float height) {
+        float offset = Math.abs(height - getHeight());
         points[2].y += offset;
         points[3].y += offset;
         size.setHeight(height);
@@ -171,7 +171,7 @@ public class Rectangle extends Polygon {
      * @see setSize
      * @see setLocation
      */
-    public void setBounds(double x, double y, double width, double height) {
+    public void setBounds(float x, float y, float width, float height) {
         setSize(width, height);
         setLocation(x, y);
     }
@@ -183,7 +183,7 @@ public class Rectangle extends Polygon {
      * @param y the new y coordinate
      * @see getLocation()
      */
-    public void setLocation(double x, double y) {
+    public void setLocation(float x, float y) {
         setX(x);
         setY(y);
     }
@@ -196,7 +196,7 @@ public class Rectangle extends Polygon {
      * @param height the specified height
      * @see getSize
      */
-    public void setSize(double width, double height) {
+    public void setSize(float width, float height) {
         setWidth(width);
         setHeight(height);
     }
@@ -213,11 +213,15 @@ public class Rectangle extends Polygon {
         setSize(dimension.width, dimension.height);
     }
 
-    //TODO
+    /**
+     * Returns the graphical center of this {@code Rectangle}.
+     * 
+     * @return the center (x, y)
+     */
     @Override
-    public Point2D.Double center() {
-        double centerX = getX() + getWidth() / 2;
-        double centerY = getY() + getHeight() / 2;
-        return new Point2D.Double(centerX, centerY);
+    public Point2D.Float center() {
+        float centerX = getX() + getWidth() / 2;
+        float centerY = getY() + getHeight() / 2;
+        return new Point2D.Float(centerX, centerY);
     }
 }
