+++ src/engine/util/Random.java
@@ -11,15 +11,15 @@ public class Random {
     private Random() {}
 
     /**
-     * Returns a double value between {@code min} and {@code max}.
+     * Returns a float value between {@code min} and {@code max}.
      * 
      * @param min the lower bound
      * @param max the higher bound
      * @return a pseudo-random between {@code min} and {@code max}
      */
-    public static double random(double min, double max) {
+    public static float random(float min, float max) {
         if (min > max)
             throw new IllegalArgumentException("boundaries must be like min <= max");
-        return min + (max - min)*rand.nextDouble();
+        return min + (max - min)*rand.nextFloat();
     }
 }
